refactor: js modules
